# Mimirtool: reduce allocations in `ParseMetricsInBoard`.