#### What this PR does
Reduce allocations in `ParseMetricsInBoard` by moving two regexps at pkg level and pre-allocating slices where possible.

To test the changes, I've copied an existing test into a benchmark.

Tested with:

`go test -run=^$ -bench=. -count=5 ./pkg/mimirtool/commands`

Benchmark comparison:

```
name                   old time/op    new time/op    delta
ParseMetricsInBoard-8    86.2µs ± 0%    80.5µs ± 0%   -6.57%  (p=0.008 n=5+5)

name                   old alloc/op   new alloc/op   delta
ParseMetricsInBoard-8    79.2kB ± 0%    67.6kB ± 0%  -14.72%  (p=0.008 n=5+5)

name                   old allocs/op  new allocs/op  delta
ParseMetricsInBoard-8     1.09k ± 0%     1.00k ± 0%   -7.82%  (p=0.008 n=5+5)
```

#### Which issue(s) this PR fixes or relates to
n.a.

#### Checklist

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`