@cristiangreco 2023-02-24
> Which problem does this solve? Looking at the benchmark I can't see any meaningful difference in absolute terms. How did you end up here?

It's used by cloud integrations to compile the list of metrics used by each one of the ~60 integrations. It's executed only once at startup of the api server, so it's a decent nice-to-have.